Move `delay` back to `src/internal/utils`
