+++ src/index.js
@@ -5,23 +5,9 @@ export { runSaga } from './internal/runSaga'
 export { END, eventChannel, channel } from './internal/channel'
 export { buffers } from './internal/buffers'
 export { takeEvery, takeLatest } from './internal/sagaHelpers'
-
-import { CANCEL } from './internal/proc'
-
-export { CANCEL }
+export { delay, CANCEL } from './internal/utils'
 
 import * as effects from './effects'
 import * as utils from './utils'
-
-export function delay(ms, val=true) {
-  let timeoutId
-  const promise = new Promise(resolve => {
-    timeoutId = setTimeout(() => resolve(val), ms)
-  })
-
-  promise[CANCEL] = () => clearTimeout(timeoutId)
-
-  return promise
-}
 
 export { effects, utils }
